requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
pinecone-client>=3.0.0
openai>=1.0.0

# Vector and embedding dependencies
sentence-transformers>=2.2.0
tiktoken>=0.5.0

# Utility dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.28.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Optional dependencies for enhanced functionality
streamlit>=1.28.0  # For web interface
gradio>=3.0.0      # Alternative web interface
